extract kodein viewmodel extensions
