+++ features-account/src/main/java/com/jbrunton/mymovies/features/account/favorites/FavoritesActivity.kt
@@ -5,7 +5,7 @@ import com.jbrunton.mymovies.features.account.R
 import com.jbrunton.mymovies.libs.ui.controllers.rootView
 import com.jbrunton.mymovies.libs.ui.livedata.observe
 import com.jbrunton.mymovies.libs.ui.views.BaseActivity
-import com.jbrunton.mymovies.libs.ui.views.injectViewModel
+import com.jbrunton.mymovies.libs.ui.viewmodels.injectViewModel
 
 class FavoritesActivity : BaseActivity<FavoritesViewModel>() {
     override val viewModel: FavoritesViewModel by injectViewModel()
